Create Prosthetic Arm
This code was for a project I worked on in undergrad in which my group created a prosthetic 3d printed arm which used an IR sensor to detect a muscle impulse which would open and close the hand. Writing the code was my responsibility alone, and I have attached the file that I wrote to my group to explain how the calibration works.